+++ src/module.ts
@@ -11,6 +11,7 @@ import {
   updateTemplates,
   addComponent,
   installModule,
+  addVitePlugin,
 } from '@nuxt/kit'
 import type { Nuxt } from '@nuxt/schema'
 import type { ModuleOptions as MDCModuleOptions } from '@nuxtjs/mdc'
@@ -24,13 +25,13 @@ import { generateCollectionInsert, generateCollectionTableDefinition } from './u
 import { componentsManifestTemplate, contentTypesTemplate, fullDatabaseRawDumpTemplate, manifestTemplate, moduleTemplates } from './utils/templates'
 import type { ResolvedCollection } from './types/collection'
 import type { ModuleOptions } from './types/module'
-import { getContentChecksum, logger, watchContents, chunks, watchComponents, startSocketServer } from './utils/dev'
+import { getContentChecksum, logger, chunks, NuxtContentHMRUnplugin } from './utils/dev'
 import { loadContentConfig } from './utils/config'
 import { createParser } from './utils/content'
 import { installMDCModule } from './utils/mdc'
 import { findPreset } from './presets'
 import type { Manifest } from './types/manifest'
-import { setupPreview, shouldEnablePreview } from './utils/preview/module'
+import { setupPreview, setupPreviewWithAPI, shouldEnablePreview } from './utils/preview/module'
 import { parseSourceBase } from './utils/source'
 import { databaseVersion, getLocalDatabase, refineDatabaseConfig, resolveDatabaseAdapter } from './utils/database'
 import type { ParsedContentFile } from './types'
@@ -53,15 +54,7 @@ export default defineNuxtModule<ModuleOptions>({
       filename: '.data/content/contents.sqlite',
     },
     preview: {},
-    watch: {
-      enabled: true,
-      port: {
-        port: 4000,
-        portRange: [4000, 4040],
-      },
-      hostname: 'localhost',
-      showURL: false,
-    },
+    watch: { enabled: true },
     renderer: {
       alias: {},
       anchorLinks: {
@@ -96,22 +89,22 @@ export default defineNuxtModule<ModuleOptions>({
     // Detect installed validators and them into content context
     await initiateValidatorsContext()
 
-    const { collections } = await loadContentConfig(nuxt)
+    const { collections } = await loadContentConfig(nuxt, options)
     manifest.collections = collections
 
-    nuxt.options.vite.optimizeDeps ||= {}
-    nuxt.options.vite.optimizeDeps.exclude ||= []
-    nuxt.options.vite.optimizeDeps.exclude.push('@sqlite.org/sqlite-wasm')
+    nuxt.options.vite.optimizeDeps = defu(nuxt.options.vite.optimizeDeps, {
+      exclude: ['@sqlite.org/sqlite-wasm'],
+    })
 
     // Ignore content directory files in building
     nuxt.options.ignore = [...(nuxt.options.ignore || []), 'content/**']
 
     // Helpers are designed to be enviroment agnostic
     addImports([
-      { name: 'queryCollection', from: resolver.resolve('./runtime/app') },
-      { name: 'queryCollectionSearchSections', from: resolver.resolve('./runtime/app') },
-      { name: 'queryCollectionNavigation', from: resolver.resolve('./runtime/app') },
-      { name: 'queryCollectionItemSurroundings', from: resolver.resolve('./runtime/app') },
+      { name: 'queryCollection', from: resolver.resolve('./runtime/client') },
+      { name: 'queryCollectionSearchSections', from: resolver.resolve('./runtime/client') },
+      { name: 'queryCollectionNavigation', from: resolver.resolve('./runtime/client') },
+      { name: 'queryCollectionItemSurroundings', from: resolver.resolve('./runtime/client') },
     ])
     addServerImports([
       { name: 'queryCollection', from: resolver.resolve('./runtime/nitro') },
@@ -122,16 +115,18 @@ export default defineNuxtModule<ModuleOptions>({
     addComponent({ name: 'ContentRenderer', filePath: resolver.resolve('./runtime/components/ContentRenderer.vue') })
 
     // Add Templates & aliases
-    nuxt.options.nitro.alias = nuxt.options.nitro.alias || {}
     addTemplate(fullDatabaseRawDumpTemplate(manifest))
-    nuxt.options.alias['#content/components'] = addTemplate(componentsManifestTemplate(manifest)).dst
-    nuxt.options.alias['#content/manifest'] = addTemplate(manifestTemplate(manifest)).dst
+    nuxt.options.alias = defu(nuxt.options.alias, {
+      '#content/components': addTemplate(componentsManifestTemplate(manifest)).dst,
+      '#content/manifest': addTemplate(manifestTemplate(manifest)).dst,
+    })
 
     // Add content types to Nuxt and Nitro
     const typesTemplateDst = addTypeTemplate(contentTypesTemplate(manifest.collections)).dst
-    nuxt.options.nitro.typescript ||= {}
-    nuxt.options.nitro.typescript.tsConfig = defu(nuxt.options.nitro.typescript.tsConfig, {
-      include: [typesTemplateDst],
+    nuxt.options.nitro.typescript = defu(nuxt.options.nitro.typescript, {
+      tsConfig: {
+        include: [typesTemplateDst],
+      },
     })
 
     // Register user components
@@ -159,29 +154,9 @@ export default defineNuxtModule<ModuleOptions>({
       }
     })
 
-    const preset = findPreset(nuxt)
-    await preset?.setup?.(options, nuxt)
-
-    // Provide default database configuration here since nuxt is merging defaults and user options
-    options.database ||= { type: 'sqlite', filename: './contents.sqlite' }
-    await refineDatabaseConfig(options._localDatabase, { rootDir: nuxt.options.rootDir, updateSqliteFileName: true })
-    await refineDatabaseConfig(options.database, { rootDir: nuxt.options.rootDir })
-
-    // Module Options
-    nuxt.options.runtimeConfig.public.content = {
-      wsUrl: '',
-    }
-    nuxt.options.runtimeConfig.content = {
-      databaseVersion,
-      version,
-      database: options.database,
-      localDatabase: options._localDatabase!,
-      integrityCheck: true,
-    } as never
-
     nuxt.hook('nitro:config', async (config) => {
       const preset = findPreset(nuxt)
-      await preset.setupNitro(config, { manifest, resolver, moduleOptions: options })
+      await preset.setupNitro(config, { manifest, resolver, moduleOptions: options, nuxt })
 
       const resolveOptions = { resolver, sqliteConnector: options.experimental?.sqliteConnector || (options.experimental?.nativeSqlite ? 'native' : undefined) }
       config.alias ||= {}
@@ -195,18 +170,42 @@ export default defineNuxtModule<ModuleOptions>({
       })
 
       // Handle HMR changes
-      if (nuxt.options.dev) {
+      if (nuxt.options.dev && options.watch?.enabled !== false) {
         addPlugin({ src: resolver.resolve('./runtime/plugins/websocket.dev'), mode: 'client' })
-        await watchComponents(nuxt)
-        const socket = await startSocketServer(nuxt, options, manifest)
-        await watchContents(nuxt, options, manifest, socket)
+        addVitePlugin(NuxtContentHMRUnplugin.vite({
+          nuxt,
+          moduleOptions: options,
+          manifest,
+        }))
       }
     })
 
     if (hasNuxtModule('nuxt-llms')) {
       installModule(resolver.resolve('./features/llms'))
     }
+
     await installMDCModule(options, nuxt)
+
+    nuxt.hook('modules:done', async () => {
+      const preset = findPreset(nuxt)
+      await preset?.setup?.(options, nuxt)
+      // Provide default database configuration here since nuxt is merging defaults and user options
+      options.database ||= { type: 'sqlite', filename: './contents.sqlite' }
+      await refineDatabaseConfig(options._localDatabase, { rootDir: nuxt.options.rootDir, updateSqliteFileName: true })
+      await refineDatabaseConfig(options.database, { rootDir: nuxt.options.rootDir })
+
+      // Module Options
+      nuxt.options.runtimeConfig.public.content = {
+        wsUrl: '',
+      }
+      nuxt.options.runtimeConfig.content = {
+        databaseVersion,
+        version,
+        database: options.database,
+        localDatabase: options._localDatabase!,
+        integrityCheck: true,
+      } as never
+    })
 
     if (nuxt.options._prepare) {
       return
@@ -233,8 +232,11 @@ export default defineNuxtModule<ModuleOptions>({
       })
 
       // Handle preview mode
-      if (shouldEnablePreview(nuxt, options)) {
+      if (hasNuxtModule('nuxt-studio')) {
         await setupPreview(options, nuxt, resolver, manifest)
+      }
+      if (shouldEnablePreview(nuxt, options)) {
+        await setupPreviewWithAPI(options, nuxt, resolver, manifest)
       }
     })
   },