+++ test/fixtures/remote-repository/nuxt.config.ts
@@ -1,0 +1,9 @@
+import { defineNuxtConfig } from 'nuxt/config'
+
+export default defineNuxtConfig({
+  modules: [
+    '../../../src/module',
+  ],
+  devtools: { enabled: true },
+  compatibilityDate: '2025-09-03',
+})