+++ CHANGELOG.md
@@ -1,6 +1,40 @@
 # Changelog
 
+## [3.7.1](https://github.com/nuxt/content/compare/v3.7.0...v3.7.1) (2025-09-16)
+
+### Bug Fixes
+
+* remove `zod` and `zod-to-json-schema` from optional deps ([#3541](https://github.com/nuxt/content/issues/3541)) ([8e038f0](https://github.com/nuxt/content/commit/8e038f0db643b05ebde15a250d0a0262eba25d2c))
+
 ## [3.7.0](https://github.com/nuxt/content/compare/v3.6.3...v3.7.0) (2025-09-12)
+
+### Deprecations
+
+The following features are deprecated and will be removed in a future release:
+- `z` re-export from `@nuxt/content`
+- Calling `.editor()` directly on zod schemas (e.g. `z.string().editor(...)`)
+
+Migration guide :
+```diff
+- import { defineContentConfig, defineCollection, z } from '@nuxt/content'
++ import { defineContentConfig, defineCollection, property } from '@nuxt/content'
++ import { z } from 'zod' // or 'zod/v3' if your setup exposes this subpath
+
+  export default defineContentConfig({
+    collections: {
+      posts: defineCollection({
+        type: 'page',
+        source: 'blog/*.md',
+        schema: z.object({
+        image: z.object({
+-         src: z.string().editor({ input: 'media' }),
++         src: property(z.string()).editor({ input: 'media' }),
+          alt: z.string(),
+        }),
+      }),
+    },
+  })
+```
 
 ### Features